chore: move verifyProjectCreated to sideBar.ts
